Rotation and zoom support in slideshow mode
drop use of elm image library - to slow and no support for jpg format
Use Browser.getElement of on images rendered off screen in order to get natural sizes of images
Combine with getViewport information to properly size images.
Allow images to be rotated and account for requreid scaling so images continue to fit on screen.
Add support for zooming in on slideshow view

default zoom and default rotation preferneces
zoom/rotation can be set on per image basis as well